@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net7.0;net8.0</TargetFrameworks>
+		<TargetFramework>net8.0</TargetFramework>
 		<Authors>Shuang Yu</Authors>
 		<Company>AhDai</Company>
 		<Version>1.0.6</Version>
@@ -43,13 +43,6 @@
 		<PackageReference Include="StackExchange.Redis" Version="2.7.17" />
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
 		<PackageReference Include="System.Drawing.Common" Version="8.0.1" />
-	</ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)' == 'net7.0' ">
-		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.15" />
-		<PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.15" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.15" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.15" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)' == 'net8.0' ">
