@@ -1,93 +1,95 @@
-﻿using AhDai.Db.Models;
+﻿using AhDai.Entity.Sys;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
 
 namespace AhDai.Db;
 
 /// <summary>
 /// DefaultDbContext
 /// </summary>
-public partial class DefaultDbContext : Microsoft.EntityFrameworkCore.DbContext
+/// <param name="options"></param>
+public partial class DefaultDbContext(DbContextOptions<DefaultDbContext> options) : Microsoft.EntityFrameworkCore.DbContext(options)
 {
-	public DefaultDbContext()
-	{
-	}
-
-	public DefaultDbContext(DbContextOptions<DefaultDbContext> options)
-		: base(options)
-	{
-	}
-
-	public virtual DbSet<Interface> Interfaces { get; set; }
-
-	public virtual DbSet<Dict> Dicts { get; set; }
-
-	public virtual DbSet<DictDatum> DictData { get; set; }
-
-	public virtual DbSet<File> Files { get; set; }
-
-	public virtual DbSet<Function> Functions { get; set; }
-
-	public virtual DbSet<FunctionInterface> FunctionInterfaces { get; set; }
-
-	public virtual DbSet<Menu> Menus { get; set; }
-
-	public virtual DbSet<OperationLog> OperationLogs { get; set; }
-
-	public virtual DbSet<Organization> Organizations { get; set; }
-
-	public virtual DbSet<Role> Roles { get; set; }
-
-	public virtual DbSet<RoleFunction> RoleFunctions { get; set; }
-
-	public virtual DbSet<User> Users { get; set; }
-
-	public virtual DbSet<UserFunction> UserFunctions { get; set; }
-
-	public virtual DbSet<UserPassword> UserPasswords { get; set; }
+    #region 系统
 
 
-	protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-	{
-		var config = Core.Utils.DbContextUtil.GetConfig();
-		optionsBuilder.UseSqlServer(config.ConnectionString, options =>
-		{
+    public virtual DbSet<Dict> Dicts { get; set; }
 
-		});
-		optionsBuilder.UseLoggerFactory(LoggerFactory.Create(builder =>
-		{
-			builder.AddConsole();
-			builder.AddDebug();
-		}));
-		optionsBuilder.AddInterceptors(new Interceptors.MySaveChangesInterceptor());
-		base.OnConfiguring(optionsBuilder);
+    public virtual DbSet<Employee> Employees { get; set; }
 
-	}
+    public virtual DbSet<File> Files { get; set; }
 
-	protected override void OnModelCreating(ModelBuilder modelBuilder)
-	{
-		modelBuilder.UseCollation("Chinese_PRC_CI_AS");
+    public virtual DbSet<Interface> Interfaces { get; set; }
 
-		modelBuilder.ApplyConfiguration(new Configuration.DictConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.DictDatumConfiguration());
+    public virtual DbSet<Menu> Menus { get; set; }
 
-		modelBuilder.ApplyConfiguration(new Configuration.OrganizationConfiguration());
+    public virtual DbSet<Organization> Organizations { get; set; }
 
-		modelBuilder.ApplyConfiguration(new Configuration.FileConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.OperationLogConfiguration());
+    public virtual DbSet<Post> Posts { get; set; }
 
-		modelBuilder.ApplyConfiguration(new Configuration.InterfaceConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.MenuConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.FunctionConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.FunctionInterfaceConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.RoleConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.RoleFunctionConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.UserConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.UserPasswordConfiguration());
-		modelBuilder.ApplyConfiguration(new Configuration.UserFunctionConfiguration());
+    public virtual DbSet<Role> Roles { get; set; }
 
-		OnModelCreatingPartial(modelBuilder);
-	}
+    public virtual DbSet<RoleMenu> RoleMenus { get; set; }
 
-	partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
+    public virtual DbSet<Tenant> Tenants { get; set; }
+
+    public virtual DbSet<TenantMenu> TenantMenus { get; set; }
+
+    public virtual DbSet<User> Users { get; set; }
+
+    public virtual DbSet<UserExtension> UserExtensions { get; set; }
+
+    public virtual DbSet<UserPassword> UserPasswords { get; set; }
+
+    public virtual DbSet<UserFunction> UserFunctions { get; set; }
+
+    public virtual DbSet<UserOrg> UserOrgs { get; set; }
+
+    public virtual DbSet<UserRole> UserRoles { get; set; }
+
+
+    public virtual DbSet<OperationLog> OperationLogs { get; set; }
+
+    #endregion
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        base.OnConfiguring(optionsBuilder);
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.UseCollation("Chinese_PRC_CI_AS");
+
+        OnSysModelCreating(modelBuilder);
+        OnModelCreatingPartial(modelBuilder);
+    }
+
+    /// <summary>
+    /// 系统
+    /// </summary>
+    /// <param name="modelBuilder"></param>
+    static void OnSysModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.DictConfiguration());
+        //modelBuilder.ApplyConfiguration(new Configuration.Sys.DictDatumConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.EmployeeConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.FileConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.InterfaceConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.MenuConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.OperationLogConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.OrganizationConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.PostConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.RoleConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.RoleMenuConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.TenantConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.TenantMenuConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.UserConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.UserExtensionConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.UserOrgConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.UserPasswordConfiguration());
+        modelBuilder.ApplyConfiguration(new Configuration.Sys.UserRoleConfiguration());
+    }
+
+
+    partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
 }
