@@ -1,17 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AhDai.Db;
+﻿namespace AhDai.Db;
 
 /// <summary>
 /// 操作用户
 /// </summary>
-/// <remarks>
-/// 构造函数
-/// </remarks>
 /// <param name="id"></param>
 /// <param name="username"></param>
 public class OperatingUser(long id, string username)
@@ -25,7 +16,19 @@ public class OperatingUser(long id, string username)
 	/// </summary>
 	public string Username { get; set; } = username;
 	/// <summary>
+	/// 姓名
+	/// </summary>
+	public string Name { get; set; } = default!;
+	/// <summary>
 	/// 租户Id
 	/// </summary>
 	public long TenantId { get; set; }
+	/// <summary>
+	/// 租户名称
+	/// </summary>
+	public string TenantName { get; set; } = "";
+	/// <summary>
+	/// 租户类型
+	/// </summary>
+	public int TenantType { get; set; }
 }
