@@ -5,12 +5,12 @@
 /// </summary>
 public class CaptchaOutput
 {
-	/// <summary>
-	/// Id
-	/// </summary>
-	public string Id { get; set; }
-	/// <summary>
-	/// 图片
-	/// </summary>
-	public string Image { get; set; }
+    /// <summary>
+    /// Id
+    /// </summary>
+    public string Id { get; set; } = "";
+    /// <summary>
+    /// 图片
+    /// </summary>
+    public string Image { get; set; } = "";
 }
