@@ -7,14 +7,14 @@ namespace AhDai.Service.Models;
 /// </summary>
 public class SignupInput
 {
-	/// <summary>
-	/// 用户名
-	/// </summary>
-	[Required]
-	public string Username { get; set; }
-	/// <summary>
-	/// 密码
-	/// </summary>
-	[Required]
-	public string Password { get; set; }
+    /// <summary>
+    /// 用户名
+    /// </summary>
+    [Required]
+    public string Username { get; set; } = "";
+    /// <summary>
+    /// 密码
+    /// </summary>
+    [Required]
+    public string Password { get; set; } = "";
 }
