@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+﻿using AhDai.WebApi.Filters;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
@@ -19,13 +20,21 @@ public static class SwaggerConfig
 	/// </summary>
 	public const string Auth = "auth";
 	/// <summary>
-	/// 业务
-	/// </summary>
-	public const string Business = "business";
-	/// <summary>
 	/// 系统
 	/// </summary>
 	public const string System = "system";
+	/// <summary>
+	/// 工作流
+	/// </summary>
+	public const string Workflow = "workflow";
+	/// <summary>
+	/// 基础
+	/// </summary>
+	public const string Basic = "basic";
+	/// <summary>
+	/// 业务
+	/// </summary>
+	public const string Business = "business";
 
 	/// <summary>
 	/// AddMySwaggerGen
@@ -33,91 +42,89 @@ public static class SwaggerConfig
 	/// <param name="services"></param>
 	/// <param name="setupAction"></param>
 	/// <returns></returns>
-	public static IServiceCollection AddMySwaggerGen(this IServiceCollection services, Action<SwaggerGenOptions> setupAction = null)
-	{
-		services.AddSwaggerGen(options =>
-		{
-			var securityScheme = new OpenApiSecurityScheme
-			{
-				Name = "JWT Authentication",
-				Description = "Enter JWT Bearer token **_only_**",
-				In = ParameterLocation.Header,
-				Type = SecuritySchemeType.Http,
-				Scheme = "bearer",
-				BearerFormat = "JWT",
-				Reference = new OpenApiReference
-				{
-					Id = JwtBearerDefaults.AuthenticationScheme,
-					Type = ReferenceType.SecurityScheme
-				}
-			};
-			options.AddSecurityDefinition(securityScheme.Reference.Id, securityScheme);
-			options.AddSecurityRequirement(new OpenApiSecurityRequirement{
-				{ securityScheme, Array.Empty<string>() }
-			});
+	public static IServiceCollection AddMySwaggerGen(this IServiceCollection services, Action<SwaggerGenOptions>? setupAction = null)
+    {
+        services.AddSwaggerGen(options =>
+        {
+            var securityScheme = new OpenApiSecurityScheme
+            {
+                Name = "JWT Authentication",
+                Description = "Enter JWT Bearer token **_only_**",
+                In = ParameterLocation.Header,
+                Type = SecuritySchemeType.Http,
+                Scheme = "bearer",
+                BearerFormat = "JWT",
+                Reference = new OpenApiReference
+                {
+                    Id = JwtBearerDefaults.AuthenticationScheme,
+                    Type = ReferenceType.SecurityScheme
+                }
+            };
+            options.AddSecurityDefinition(securityScheme.Reference.Id, securityScheme);
+            options.AddSecurityRequirement(new OpenApiSecurityRequirement{
+                { securityScheme, Array.Empty<string>() }
+            });
 
-			var terms = new Uri("https://example.com/terms");
-			var contact = new OpenApiContact
-			{
-				Name = "联系人",
-				Url = new Uri("https://example.com/contact"),
-				Email = "example@live.com"
-			};
-			var license = new OpenApiLicense
-			{
-				Name = "Example License",
-				Url = new Uri("https://example.com/license")
-			};
-			options.SwaggerDoc(Auth, new OpenApiInfo { Title = "认证", Description = "by AhDai", Version = "v1", TermsOfService = terms, Contact = contact, License = license });
-			options.SwaggerDoc(Business, new OpenApiInfo { Title = "业务", Description = "by AhDai", Version = "v1" });
-			options.SwaggerDoc(System, new OpenApiInfo { Title = "系统", Description = "by AhDai", Version = "v1" });
-			options.DocInclusionPredicate((docName, apiDesc) =>
-			{
-				return true;
-			});
+            var terms = new Uri("https://example.com/terms");
+            var contact = new OpenApiContact
+            {
+                Name = "联系人",
+                Url = new Uri("https://example.com/contact"),
+                Email = "example@live.com"
+            };
+            var license = new OpenApiLicense
+            {
+                Name = "Example License",
+                Url = new Uri("https://example.com/license")
+            };
+            options.SwaggerDoc(Business, new OpenApiInfo { Title = "业务", Description = "by AhDai", Version = "v1" });
+            options.SwaggerDoc(System, new OpenApiInfo { Title = "系统", Description = "by AhDai", Version = "v1" });
+            options.DocInclusionPredicate((docName, apiDesc) =>
+            {
+                return true;
+            });
 
-			var baseDirectory = AppContext.BaseDirectory;
-			var xmlPaths = new string[] {
-				Path.Combine(baseDirectory, "AhDai.Core.xml"),
-				Path.Combine(baseDirectory, "AhDai.Service.xml")
-			};
-			foreach (var xmlPath in xmlPaths)
-			{
-				if (!Path.Exists(xmlPath)) continue;
-				options.IncludeXmlComments(xmlPath);
-			}
-			options.IncludeXmlComments(Path.Combine(baseDirectory, $"{AppDomain.CurrentDomain.FriendlyName}.xml"), true);
+            var baseDirectory = AppContext.BaseDirectory;
+            var xmlPaths = new string[] {
+                Path.Combine(baseDirectory, "AhDai.Core.xml"),
+                Path.Combine(baseDirectory, "AhDai.Service.xml")
+            };
+            foreach (var xmlPath in xmlPaths)
+            {
+                if (!Path.Exists(xmlPath)) continue;
+                options.IncludeXmlComments(xmlPath);
+            }
+            options.IncludeXmlComments(Path.Combine(baseDirectory, $"{AppDomain.CurrentDomain.FriendlyName}.xml"), true);
 
-			options.OperationFilter<Filters.SwaggerOperationFilter>();
-			options.DocumentFilter<Filters.SwaggerDocumentFilter>();
+            options.OperationFilter<SwaggerOperationFilter>();
+            options.DocumentFilter<SwaggerDocumentFilter>();
 
-			setupAction?.Invoke(options);
-		});
+            setupAction?.Invoke(options);
+        });
 
-		return services;
-	}
+        return services;
+    }
 
-	/// <summary>
-	/// UseMySwaggerUI
-	/// </summary>
-	/// <param name="app"></param>
-	/// <param name="setupAction"></param>
-	/// <returns></returns>
-	public static IApplicationBuilder UseMySwaggerUI(this IApplicationBuilder app, Action<SwaggerUIOptions> setupAction = null)
-	{
-		app.UseSwaggerUI(options =>
-		{
-			var root = "";
-			options.SwaggerEndpoint($"{root}/swagger/{Auth}/swagger.json", "认证");
-			options.SwaggerEndpoint($"{root}/swagger/{Business}/swagger.json", "业务");
-			options.SwaggerEndpoint($"{root}/swagger/{System}/swagger.json", "系统");
-			options.ShowExtensions();
-			//options.RoutePrefix = string.Empty;
+    /// <summary>
+    /// UseMySwaggerUI
+    /// </summary>
+    /// <param name="app"></param>
+    /// <param name="setupAction"></param>
+    /// <returns></returns>
+    public static IApplicationBuilder UseMySwaggerUI(this IApplicationBuilder app, Action<SwaggerUIOptions>? setupAction = null)
+    {
+        app.UseSwaggerUI(options =>
+        {
+            var root = "";
+            options.SwaggerEndpoint($"{root}/swagger/{Business}/swagger.json", "业务");
+            options.SwaggerEndpoint($"{root}/swagger/{System}/swagger.json", "系统");
+            options.ShowExtensions();
+            //options.RoutePrefix = string.Empty;
 
-			setupAction?.Invoke(options);
-		});
+            setupAction?.Invoke(options);
+        });
 
-		return app;
-	}
+        return app;
+    }
 
 }
