@@ -13,81 +13,66 @@ namespace AhDai.WebApi.Controllers;
 /// </summary>
 /// <param name="service"></param>
 [ApiExplorerSettings(GroupName = Configs.SwaggerConfig.Auth)]
-[Route("api/auth")]
+[Route("api/v1/auth")]
 public class AuthController(IAuthService service) : ControllerBase
 {
-	readonly IAuthService _service = service;
+    readonly IAuthService _service = service;
 
-	/// <summary>
-	/// 获取验证码
-	/// </summary>
-	/// <returns></returns>
-	[HttpGet("captcha")]
-	public async Task<ApiResult<CaptchaOutput>> CaptchaAsync()
-	{
-		var res = await _service.GenerateCaptchaAsync();
-		return ApiResult.Success(res);
-	}
+    /// <summary>
+    /// 注册
+    /// </summary>
+    /// <returns></returns>
+    [AllowAnonymous]
+    [HttpPost("register")]
+    public async Task<IApiResult<string>> RegisterAsync([FromBody] SignupInput input)
+    {
+        await _service.RegisterAsync(input);
+        return ApiResult.Success();
+    }
 
-	/// <summary>
-	/// 获取登录信息
-	/// </summary>
-	/// <returns></returns>
-	[Authorize]
-	[HttpGet("login")]
-	public ApiResult<LoginResult> GetLogin()
-	{
-		var token = HttpContext.Request.Headers[Const.Authorization].FirstOrDefault();
-		var res = _service.GetLoginAsync(token);
-		return ApiResult.Success(res);
-	}
+    /// <summary>
+    /// 获取登录信息
+    /// </summary>
+    /// <returns></returns>
+    [HttpGet("login")]
+    public async Task<IApiResult<LoginData>> GetLoginAsync()
+    {
+        var res = await _service.GetLoginAsync();
+        return ApiResult.Success(res);
+    }
 
-	/// <summary>
-	/// 登录
-	/// </summary>
-	/// <param name="input">入参</param>
-	/// <returns></returns>
-	[HttpPost("login")]
-	public async Task<ApiResult<LoginResult>> LoginAsync([FromBody] LoginInput input)
-	{
-		var token = await _service.LoginAsync(input);
-		return ApiResult.Success(token);
-	}
+    /// <summary>
+    /// 登录
+    /// </summary>
+    /// <param name="input">入参</param>
+    /// <returns></returns>
+    [AllowAnonymous]
+    [HttpPost("login")]
+    public async Task<IApiResult<LoginResult>> LoginAsync([FromBody] LoginInput input)
+    {
+        var token = await _service.LoginAsync(input);
+        return ApiResult.Success(token);
+    }
 
-	/// <summary>
-	/// 登出
-	/// </summary>
-	/// <returns></returns>
-	[HttpDelete("logout")]
-	public async Task<ApiResult<bool>> LogoutAsync()
-	{
-		var token = HttpContext.Request.Headers[Const.Authorization].FirstOrDefault();
-		var res = await _service.LogoutAsync(token);
-		return ApiResult.Success(res);
-	}
+    /// <summary>
+    /// 登出
+    /// </summary>
+    /// <returns></returns>
+    [HttpDelete("logout")]
+    public async Task<IApiResult<bool>> LogoutAsync()
+    {
+        var res = await _service.LogoutAsync();
+        return ApiResult.Success(res);
+    }
 
-	/// <summary>
-	/// 刷新Token
-	/// </summary>
-	/// <returns></returns>
-	[Authorize]
-	[HttpPut("refreshToken")]
-	public async Task<ApiResult<LoginResult>> RefreshTokenAsync()
-	{
-		var token = HttpContext.Request.Headers[Const.Authorization].FirstOrDefault();
-		var res = await _service.RefreshTokenAsync(token);
-		return ApiResult.Success(res);
-	}
-
-	/// <summary>
-	/// 注册
-	/// </summary>
-	/// <returns></returns>
-	[Authorize]
-	[HttpPost("signup")]
-	public async Task<ApiResult<string>> SignupAsync([FromBody] SignupInput input)
-	{
-		await _service.SignupAsync(input);
-		return ApiResult.Success();
-	}
+    /// <summary>
+    /// 刷新Token
+    /// </summary>
+    /// <returns></returns>
+    [HttpPut("refreshToken")]
+    public async Task<IApiResult<LoginResult>> RefreshTokenAsync()
+    {
+        var res = await _service.RefreshTokenAsync();
+        return ApiResult.Success(res);
+    }
 }
